review

+++ src/internal/signalr/handler.go
@@ -150,7 +150,7 @@ func (h *observationsHandler) handleMaxChargerCurrent(observation model.Observat
 		return err
 	}
 
-	ok := h.cache.SetMaxCurrent(int64(math.Round(val)), observation.Timestamp)
+	ok := h.cache.SetMaxCurrent(int(math.Round(val)), observation.Timestamp)
 	if !ok {
 		return nil
 	}
@@ -186,7 +186,7 @@ func (h *observationsHandler) handleDynamicChargerCurrent(observation model.Obse
 		return err
 	}
 
-	ok := h.cache.SetOfferedCurrent(int64(math.Round(val)), observation.Timestamp)
+	ok := h.cache.SetOfferedCurrent(int(math.Round(val)), observation.Timestamp)
 	if !ok {
 		return nil
 	}
@@ -228,7 +228,7 @@ func (h *observationsHandler) handleCableRating(observation model.Observation) e
 		return err
 	}
 
-	cableCurrent := int64(val)
+	cableCurrent := int(val)
 
 	ok := h.cache.SetCableCurrent(cableCurrent, observation.Timestamp)
 	if !ok {
